Clean up gatherxitinputnodehandler
