+++ src/gatherexitnodeinputhandler.hpp
@@ -21,22 +21,13 @@
 #include <openvino/openvino.hpp>
 
 #include "gathernodeinputhandler.hpp"
-
-#pragma GCC diagnostic push
-#pragma GCC diagnostic ignored "-Wall"
-#include "tensorflow/core/framework/tensor.h"
-#include "tensorflow_serving/apis/prediction_service.grpc.pb.h"
-#pragma GCC diagnostic pop
-
 #include "kfs_frontend/kfs_utils.hpp"
-#include "kfs_grpc_inference_service.hpp"
 #include "logging.hpp"
 #include "profiler.hpp"
 #include "status.hpp"
+#include "tfs_frontend/tfs_utils.hpp"
 
 namespace ovms {
-
-Status prepareConsolidatedTensorImpl(tensorflow::serving::PredictResponse* response, char*& tensorOut, const std::string& name, size_t size);
 
 template <class ResponseType>
 class GatherExitNodeInputHandler : public GatherNodeInputHandler {
