+++ src/tfs_frontend/tfs_utils.cpp
@@ -20,8 +20,11 @@
 #include <sstream>
 #include <string>
 #include <unordered_map>
+#include <utility>
 
 #include "../logging.hpp"
+#include "../profiler.hpp"
+#include "../status.hpp"
 
 namespace ovms {
 
@@ -115,4 +118,16 @@ Precision TFSPrecisionToOvmsPrecision(const TFSDataType& datatype) {
     return it->second;
 }
 
+Status prepareConsolidatedTensorImpl(tensorflow::serving::PredictResponse* response, char*& bufferOut, const std::string& name, size_t size) {
+    OVMS_PROFILE_FUNCTION();
+    tensorflow::TensorProto tensorProto;
+    auto [it, isInserted] = response->mutable_outputs()->insert(google::protobuf::MapPair<std::string, tensorflow::TensorProto>(name, std::move(tensorProto)));
+    if (!isInserted) {
+        SPDLOG_LOGGER_ERROR(dag_executor_logger, "Failed to prepare consolidated tensor, tensor with name {} already prepared", name);
+        return StatusCode::INTERNAL_ERROR;
+    }
+    it->second.mutable_tensor_content()->resize(size);
+    bufferOut = it->second.mutable_tensor_content()->data();
+    return StatusCode::OK;
+}
 }  // namespace ovms
